code spliting of mapScreen
