+++ navigation/MapStackParamList.ts
@@ -1,3 +1,4 @@
+import { NativeStackScreenProps } from "@react-navigation/native-stack";
 export type MapStackParamList = {
   Login: undefined;
   SignUp: undefined;
