--- utils/Status.tsx
@@ -1,29 +1,0 @@
-// export enum Status {
-//   READY = "Ready",
-//   BUSY = "Busy",
-//   CLOSED = "CLOSED",
-// }
-
-// const statusValues = Object.values(Status);
-// export const statusColorMapping = {
-//   [Status.READY]: "#0ECC6D",
-//   [Status.BUSY]: "#ff6b06",
-//   [Status.CLOSED]: "#d71515",
-// };
-
-// export const getRandomStatus = () => {
-//   const randomIndex = Math.floor(Math.random() * statusValues.length);
-//   return statusValues[randomIndex];
-// };
-
-// export const getStatusColor = (status: Status) => {
-//   return statusColorMapping[status];
-// };
-
-// const RandomStatus = () => {
-//   const status = getRandomStatus();
-//   console.log(status);
-//   return status;
-// };
-
-// export default RandomStatus;
